car-diag/requirements.txt
# car-diag dependencies

# Serial communication
pyserial>=3.5,<4.0

# GUI
PyQt6>=6.5,<7.0
pyqtgraph>=0.13,<1.0

# Development / Testing
pytest>=8.0,<9.0
pytest-qt>=4.2,<5.0
